added input to check before resetting

+++ update_items_db.py
@@ -50,46 +50,47 @@ implicitModifiers
 explicitModifiers
 aliases
 '''
+if input("Do you want to delete and recollect all item collections? y/n") == "y":
+    collection = poe_client["unique_weapons"]
+    collection.drop()
+    collection = poe_client["divination_cards"]
+    collection.drop()
+    collection = poe_client["unique_armours"]
+    collection.drop()
+    collection = poe_client["unique_accessories"]
+    collection.drop()
+    collection = poe_client["unique_jewels"]
+    collection.drop()
+    collection = poe_client["unique_flasks"]
+    collection.drop()
 
-collection = poe_client["unique_weapons"]
-collection.drop()
-collection = poe_client["divination_cards"]
-collection.drop()
-collection = poe_client["unique_armours"]
-collection.drop()
-collection = poe_client["unique_accessories"]
-collection.drop()
-collection = poe_client["unique_jewels"]
-collection.drop()
-collection = poe_client["unique_flasks"]
-collection.drop()
+    collection = poe_client["unique_weapons"]
+    collection = poe_client["divination_cards"]
+    collection = poe_client["unique_armours"]
+    collection = poe_client["unique_accessories"]
+    collection = poe_client["unique_jewels"]
+    collection = poe_client["unique_flasks"]
 
-collection = poe_client["unique_weapons"]
-collection = poe_client["divination_cards"]
-collection = poe_client["unique_armours"]
-collection = poe_client["unique_accessories"]
-collection = poe_client["unique_jewels"]
-collection = poe_client["unique_flasks"]
-
-for item_type in test_type_routes:
-    request_string = 'https://poe.ninja/api/data/itemoverview?league={}&type={}'.format(current_league, item_type)
-    r = requests.get(request_string)
-    specific_type_collection = poe_client.get_collection(item_type_to_db[item_type])
-    for item in r.json()['lines']:
-        item_to_add = {}
-        item_to_add['id'] = item['id']
-        item_to_add['name'] = item['name']
-        item_to_add['itemType'] = item_type
-        item_to_add['baseType'] = item['baseType']
-        item_to_add['levelRequired'] = item["levelRequired"]
-        item_to_add['itemType'] = item['itemType']
-        item_to_add['icon'] = item["icon"]
-        item_to_add['baseType'] = item["baseType"]
-        item_to_add['aliases'] = [item['name'].replace("'", "").lower()]
-        item_to_add['implicitModifiers'] = [item["implicitModifiers"][i]['text'] for i in range(len(item["implicitModifiers"]))]
-        item_to_add['explicitModifiers'] = [item["explicitModifiers"][i]['text'] for i in range(len(item["explicitModifiers"]))]
-        item_to_add['url'] = "https://pathofexile.gamepedia.com/" + item['name'].replace(" ","_")
-        print(item_to_add)
-        specific_type_collection.insert_one(item_to_add)
+    for item_type in test_type_routes:
+        request_string = 'https://poe.ninja/api/data/itemoverview?league={}&type={}'.format(current_league, item_type)
+        r = requests.get(request_string)
+        specific_type_collection = poe_client.get_collection(item_type_to_db[item_type])
+        for item in r.json()['lines']:
+            item_to_add = {}
+            item_to_add['id'] = item['id']
+            item_to_add['name'] = item['name']
+            item_to_add['itemCategory'] = item_type
+            item_to_add['itemBase'] = item['itemType']
+            item_to_add['baseType'] = item['baseType']
+            item_to_add['levelRequired'] = item["levelRequired"]
+            item_to_add['icon'] = item["icon"]
+            item_to_add['baseType'] = item["baseType"]
+            item_to_add['aliases'] = [item['name'].replace("'", "").lower()]
+            item_to_add['implicitModifiers'] = [item["implicitModifiers"][i]['text'] for i in range(len(item["implicitModifiers"]))]
+            item_to_add['explicitModifiers'] = [item["explicitModifiers"][i]['text'] for i in range(len(item["explicitModifiers"]))]
+            item_to_add['url'] = "https://pathofexile.gamepedia.com/" + item['name'].replace(" ","_")
+            item_to_add['flavourText'] = item['flavourText']
+            print(item_to_add)
+            specific_type_collection.insert_one(item_to_add)
